Struct treap_map::Node

pub struct Node<C: TreapMapConfig> {
    pub key: C::SearchKey,
    pub value: C::Value,
    pub sort_key: C::SortKey,
    pub weight: C::Weight,
    /* private fields */
}

A node in a treap-map data structure.

The Node struct represents a node in a treap-map and contains various key-value pairs and metadata required for the proper functioning and maintenance of the treap-map. Direct modification of these fields is not recommended outside of the TreapMap::update function, as this function correctly maintains the integrity of the treap-map.

Fields
key: C::SearchKey

The key exposed externally. Used for key-based searches within the treap-map.

value: C::Value

The value stored in the node.

sort_key: C::SortKey

The sorting key for the treap-map. If the type is (), the search_key is used for sorting.

weight: C::Weight

The weight of the node, used by the treap-map to maintain accumulated weights.

Implementations

impl<C: TreapMapConfig> Node<C>

pub fn new( key: C::SearchKey, value: C::Value, sort_key: C::SortKey, weight: C::Weight, priority: u64 ) -> Node<C>

pub fn sum_weight(&self) -> C::Weight