Crate treap_map

Structs

ApplyOpOutcome

Represents the outcome of an operation applied in the TreapMap::update function.

Iter

NoWeight

Represents a dummy version of ConsolidatableWeight.

Node

A node in a treap-map data structure.

TreapMap

A treap map data structure.

Enums

Direction

Represents the possible directions in a binary tree search based on key comparisons.

SearchDirection

Represents the directions for the search in accumulate_weight_search.

SearchResult

Represents the possible outcomes of the accumulate_weight_search.

Traits

ConsoliableWeight

The weight type in a Treap. It is used to perform operations like calculating sums or maximum values of an interval in logrithmic time over treap.

KeyMngTrait

Searching in Treap requires sort key. This trait manages the relationship among sort keys, search keys and values in a Treap. This is necessary when the sort key is not directly derivable from the search key or is not a null element.

SharedKeyTreapMapConfig

If TreapMap is indexed in the same key as the inside treap, it can be configed in a simple way.

TreapMapConfig

TreapMap is a struct which implements a treap which can be indexed by a different key (type SearchKey). The associate type SortKey and SearchKey defines how to order node in treap collaborately.

Functions

accumulate_weight_search

Performs a binary search in a treap-map.