Trait treap_map::SharedKeyTreapMapConfig

pub trait SharedKeyTreapMapConfig {
    type Key: Ord;
    type Value: Clone;
    type Weight: ConsoliableWeight;
}

If TreapMap is indexed in the same key as the inside treap, it can be configed in a simple way.

Required Associated Types

type Key: Ord

The search key in the TreapMap.

type Value: Clone

The stored value.

type Weight: ConsoliableWeight

The consolidable weight.